+++ src/backEnd/app.js
@@ -12,20 +12,23 @@ let corsPort = process.env.NODE_ENV === "development" ? "3000":"5000";
 
 var app = express();
 
-//--------------production mode 下把 build 的東西掛到 express server 上 ---------------------
-app.use(express.static(path.join(__dirname,'../../build')));
-  
-app.get('/',function(req,res){
-    res.sendFile(path.join(__dirname,'../../build','index.html'));
-  })
-//-----------------------------------------------------------------------------------------
-
 app.use(cors({credentials:true, origin:`http://localhost:${corsPort}`})); // option credentials為解決cookie跨域不可傳遞之問題
 
 app.use(cookieParser());
 // app.use(express.static('public')
 
 app.use('/rest', router);
+
+//--------------production mode 下把 build 的東西掛到 express server 上 ---------------------
+app.use(express.static(path.join(__dirname,'../../build')));
+  
+app.get(/\/[^rest]/,function(req,res){
+    if(!req.cookies.jwtToken){  //若未登入(無 jwtToken) 則導到登入頁 , 避免使用者未登入直接輸入網址進入功能頁面
+      res.redirect('/');
+    }
+    res.sendFile(path.join(__dirname,'../../build','index.html'));
+  })
+//-----------------------------------------------------------------------------------------
 
 
 app.listen(serverPort, function () {
