+++ src/index.js
@@ -1,42 +1,59 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-// import './index.css';
-// import App from './App';
-// import Counter from './counter/counter'
-// import Transformation from './transFormation/transFormation';
 import * as serviceWorker from './serviceWorker';
 import { Provider } from 'react-redux';
 import {store} from './store/store.js'
-import {Navbar} from './module/navBar/navBar.jsx';
+import {Navbar} from '../src/module/navBar/navBar.jsx';
+import {IncreaseForm} from '../src/module/nameList/IncreaseNameList.jsx'
+import {DeleteForm} from '../src/module/nameList/DeleteNameList.jsx'
+import {UpdateForm} from '../src/module/nameList/UpdateNameList.jsx'
+import {Home} from '../src/module/homePage/home.js'
+import {Login} from '../src/module/Login/login.jsx'
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  useHistory,
+  useRouteMatch
+} from "react-router-dom";
 
-// const HomeBlock = styled.div`
-//     background-color:pink;
-// `;
 
-
-// const HomeTitle = styled.div`
-//     font-size:50px;
-//     font-weight:bold;
-//     display:flex;
-//     items-align:center;
-//     justify-content:center;
-// `;
-
-// const Home = ()=>{
-//   return (
-//     <>
-//   </>
-//   )
-// }
 
 ReactDOM.render(
 
   // <React.StrictMode>
   <Provider store={store}>
-    {/* {counters.map((eachElement)=>(<Counter/>))} */}
-    {/* <Transformation/>   */}
-  <Navbar/>
-  {/* <Component2/> */}
+            <Router>
+                  <Switch>
+                      <Route exact path="/">
+                                    <Login/>
+                      </Route>
+                      <Route exact path="/Home">
+                                <>
+                                    <Navbar/>
+                                    <Home/>
+                                </>
+                      </Route>
+                      <Route exact path="/Increase">
+                                <>
+                                    <Navbar/>
+                                    <IncreaseForm/>
+                                </>
+                      </Route>
+                      <Route path="/Delete">
+                                <>
+                                    <Navbar/>
+                                    <DeleteForm/>
+                                </>
+                      </Route>
+                      <Route path="/Update">
+                                <>
+                                    <Navbar/>
+                                    <UpdateForm/>
+                                </>
+                      </Route>
+                  </Switch>
+            </Router>
   </Provider>,
   // </React.StrictMode>,
   document.getElementById('root')
