+++ src/module/Login/login.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from '@emotion/styled'
 import Input from './input.jsx'
-import {useSelector} from 'react-redux';
 import {useHistory} from "react-router-dom";
 import { useEffect } from 'react';
 import {getCookie} from '../../common/cookie.js'
@@ -20,7 +19,7 @@ const Content = styled.div`
 `;
 
 export function Login() {
-  let isLogin = getCookie("jwtToken")===undefined?false:true;
+  let isLogin = getCookie("jwtToken")===undefined?false:true; 
   const history = useHistory();
   const mayPushToHome = ()=>{   //如果為登入狀態，使用者在另個視窗開啟 /url 則會導到 home 頁 
         if(isLogin){
